+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2025-10-24
+
+### Added
+- changed namespace to `runningIn`
+-   `getInfo()` function that provides detailed environment information including OS, browser, device, capabilities, and runtime details.
+-   Dynamic parsing of user agent strings for browser environments.
+-   System information extraction for Node.js environments
+
 ## [1.0.0] - 2025-09-18
 
 ### Added